src/browser-agent/requirements.txt
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
langchain-google-genai>=0.0.6
python-dotenv>=1.0.0
browser-use>=0.4.0 